move constants out of client
