+++ constants/locales.js
